+++ crates/icydb-core/src/db/executor/prepared_execution_plan/shared_plan.rs
@@ -120,9 +120,7 @@ impl SharedPreparedExecutionPlan {
         self.core.residents.index_prefix_specs.as_ref()
     }
 
-    pub(in crate::db::executor) fn index_range_specs(
-        &self,
-    ) -> &[crate::db::executor::LoweredIndexRangeSpec] {
+    pub(in crate::db) fn index_range_specs(&self) -> &[crate::db::executor::LoweredIndexRangeSpec] {
         self.core.residents.index_range_specs.as_ref()
     }